Finalização do projeto de hackathon

+++ hackaton/Models/Usuario.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string Nome { get; set; }
-        public float PesoEficiencia { get; set; }
+        public double PesoEficiencia { get; set; }
         public IList<RelatorioHora> ListaRelatorioHora { get; set; }
     }
 }
